[android] Build CoreGraphics

+++ CoreGraphics/CGContext.h
@@ -107,17 +107,17 @@ COREGRAPHICS_EXPORT bool CGContextPathContainsPoint(CGContextRef context, CGPoin
 
 COREGRAPHICS_EXPORT void CGContextBeginPath(CGContextRef context);
 COREGRAPHICS_EXPORT void CGContextClosePath(CGContextRef context);
-COREGRAPHICS_EXPORT void CGContextMoveToPoint(CGContextRef context, float x, float y);
-COREGRAPHICS_EXPORT void CGContextAddLineToPoint(CGContextRef context, float x, float y);
-COREGRAPHICS_EXPORT void CGContextAddCurveToPoint(CGContextRef context, float cx1, float cy1, float cx2, float cy2, float x, float y);
-COREGRAPHICS_EXPORT void CGContextAddQuadCurveToPoint(CGContextRef context, float cx1, float cy1, float x, float y);
+COREGRAPHICS_EXPORT void CGContextMoveToPoint(CGContextRef context, CGFloat x, CGFloat y);
+COREGRAPHICS_EXPORT void CGContextAddLineToPoint(CGContextRef context, CGFloat x, CGFloat y);
+COREGRAPHICS_EXPORT void CGContextAddCurveToPoint(CGContextRef context, CGFloat cx1, CGFloat cy1, CGFloat cx2, CGFloat cy2, CGFloat x, CGFloat y);
+COREGRAPHICS_EXPORT void CGContextAddQuadCurveToPoint(CGContextRef context, CGFloat cx1, CGFloat cy1, CGFloat x, CGFloat y);
 
 COREGRAPHICS_EXPORT void CGContextAddLines(CGContextRef context, const CGPoint *points, unsigned count);
 COREGRAPHICS_EXPORT void CGContextAddRect(CGContextRef context, CGRect rect);
 COREGRAPHICS_EXPORT void CGContextAddRects(CGContextRef context, const CGRect *rects, unsigned count);
 
-COREGRAPHICS_EXPORT void CGContextAddArc(CGContextRef context, float x, float y, float radius, float startRadian, float endRadian, bool clockwise);
-COREGRAPHICS_EXPORT void CGContextAddArcToPoint(CGContextRef context, float x1, float y1, float x2, float y2, float radius);
+COREGRAPHICS_EXPORT void CGContextAddArc(CGContextRef context, CGFloat x, CGFloat y, CGFloat radius, CGFloat startRadian, CGFloat endRadian, bool clockwise);
+COREGRAPHICS_EXPORT void CGContextAddArcToPoint(CGContextRef context, CGFloat x1, CGFloat y1, CGFloat x2, CGFloat y2, CGFloat radius);
 COREGRAPHICS_EXPORT void CGContextAddEllipseInRect(CGContextRef context, CGRect rect);
 
 COREGRAPHICS_EXPORT void CGContextAddPath(CGContextRef context, CGPathRef path);
@@ -145,9 +145,9 @@ COREGRAPHICS_EXPORT CGRect CGContextConvertRectToDeviceSpace(CGContextRef contex
 COREGRAPHICS_EXPORT CGRect CGContextConvertRectToUserSpace(CGContextRef context, CGRect rect);
 
 COREGRAPHICS_EXPORT void CGContextConcatCTM(CGContextRef context, CGAffineTransform matrix);
-COREGRAPHICS_EXPORT void CGContextTranslateCTM(CGContextRef context, float translatex, float translatey);
-COREGRAPHICS_EXPORT void CGContextScaleCTM(CGContextRef context, float scalex, float scaley);
-COREGRAPHICS_EXPORT void CGContextRotateCTM(CGContextRef context, float radians);
+COREGRAPHICS_EXPORT void CGContextTranslateCTM(CGContextRef context, CGFloat translatex, CGFloat translatey);
+COREGRAPHICS_EXPORT void CGContextScaleCTM(CGContextRef context, CGFloat scalex, CGFloat scaley);
+COREGRAPHICS_EXPORT void CGContextRotateCTM(CGContextRef context, CGFloat radians);
 
 COREGRAPHICS_EXPORT void CGContextClip(CGContextRef context);
 COREGRAPHICS_EXPORT void CGContextEOClip(CGContextRef context);
@@ -158,54 +158,54 @@ COREGRAPHICS_EXPORT void CGContextClipToRects(CGContextRef context, const CGRect
 COREGRAPHICS_EXPORT void CGContextSetStrokeColorSpace(CGContextRef context, CGColorSpaceRef colorSpace);
 COREGRAPHICS_EXPORT void CGContextSetFillColorSpace(CGContextRef context, CGColorSpaceRef colorSpace);
 
-COREGRAPHICS_EXPORT void CGContextSetStrokeColor(CGContextRef context, const float *components);
+COREGRAPHICS_EXPORT void CGContextSetStrokeColor(CGContextRef context, const CGFloat *components);
 COREGRAPHICS_EXPORT void CGContextSetStrokeColorWithColor(CGContextRef context, CGColorRef color);
-COREGRAPHICS_EXPORT void CGContextSetGrayStrokeColor(CGContextRef context, float gray, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetRGBStrokeColor(CGContextRef context, float r, float g, float b, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetCMYKStrokeColor(CGContextRef context, float c, float m, float y, float k, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetCalibratedRGBStrokeColor(CGContextRef context, float red, float green, float blue, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetCalibratedGrayStrokeColor(CGContextRef context, float gray, float alpha);
+COREGRAPHICS_EXPORT void CGContextSetGrayStrokeColor(CGContextRef context, CGFloat gray, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetRGBStrokeColor(CGContextRef context, CGFloat r, CGFloat g, CGFloat b, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetCMYKStrokeColor(CGContextRef context, CGFloat c, CGFloat m, CGFloat y, CGFloat k, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetCalibratedRGBStrokeColor(CGContextRef context, CGFloat red, CGFloat green, CGFloat blue, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetCalibratedGrayStrokeColor(CGContextRef context, CGFloat gray, CGFloat alpha);
 
-COREGRAPHICS_EXPORT void CGContextSetFillColor(CGContextRef context, const float *components);
+COREGRAPHICS_EXPORT void CGContextSetFillColor(CGContextRef context, const CGFloat *components);
 COREGRAPHICS_EXPORT void CGContextSetFillColorWithColor(CGContextRef context, CGColorRef color);
-COREGRAPHICS_EXPORT void CGContextSetGrayFillColor(CGContextRef context, float gray, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetRGBFillColor(CGContextRef context, float r, float g, float b, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetCMYKFillColor(CGContextRef context, float c, float m, float y, float k, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetCalibratedGrayFillColor(CGContextRef context, float gray, float alpha);
-COREGRAPHICS_EXPORT void CGContextSetCalibratedRGBFillColor(CGContextRef context, float red, float green, float blue, float alpha);
+COREGRAPHICS_EXPORT void CGContextSetGrayFillColor(CGContextRef context, CGFloat gray, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetRGBFillColor(CGContextRef context, CGFloat r, CGFloat g, CGFloat b, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetCMYKFillColor(CGContextRef context, CGFloat c, CGFloat m, CGFloat y, CGFloat k, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetCalibratedGrayFillColor(CGContextRef context, CGFloat gray, CGFloat alpha);
+COREGRAPHICS_EXPORT void CGContextSetCalibratedRGBFillColor(CGContextRef context, CGFloat red, CGFloat green, CGFloat blue, CGFloat alpha);
 
-COREGRAPHICS_EXPORT void CGContextSetAlpha(CGContextRef context, float alpha);
+COREGRAPHICS_EXPORT void CGContextSetAlpha(CGContextRef context, CGFloat alpha);
 
 COREGRAPHICS_EXPORT void CGContextSetPatternPhase(CGContextRef context, CGSize phase);
-COREGRAPHICS_EXPORT void CGContextSetStrokePattern(CGContextRef context, CGPatternRef pattern, const float *components);
-COREGRAPHICS_EXPORT void CGContextSetFillPattern(CGContextRef context, CGPatternRef pattern, const float *components);
+COREGRAPHICS_EXPORT void CGContextSetStrokePattern(CGContextRef context, CGPatternRef pattern, const CGFloat *components);
+COREGRAPHICS_EXPORT void CGContextSetFillPattern(CGContextRef context, CGPatternRef pattern, const CGFloat *components);
 
 COREGRAPHICS_EXPORT void CGContextSetTextMatrix(CGContextRef context, CGAffineTransform matrix);
 
-COREGRAPHICS_EXPORT void CGContextSetTextPosition(CGContextRef context, float x, float y);
-COREGRAPHICS_EXPORT void CGContextSetCharacterSpacing(CGContextRef context, float spacing);
+COREGRAPHICS_EXPORT void CGContextSetTextPosition(CGContextRef context, CGFloat x, CGFloat y);
+COREGRAPHICS_EXPORT void CGContextSetCharacterSpacing(CGContextRef context, CGFloat spacing);
 COREGRAPHICS_EXPORT void CGContextSetTextDrawingMode(CGContextRef context, CGTextDrawingMode textMode);
 
 COREGRAPHICS_EXPORT void CGContextSetFont(CGContextRef context, CGFontRef font);
-COREGRAPHICS_EXPORT void CGContextSetFontSize(CGContextRef context, float size);
-COREGRAPHICS_EXPORT void CGContextSelectFont(CGContextRef context, const char *name, float size, CGTextEncoding encoding);
+COREGRAPHICS_EXPORT void CGContextSetFontSize(CGContextRef context, CGFloat size);
+COREGRAPHICS_EXPORT void CGContextSelectFont(CGContextRef context, const char *name, CGFloat size, CGTextEncoding encoding);
 COREGRAPHICS_EXPORT void CGContextSetShouldSmoothFonts(CGContextRef context, bool yesOrNo);
 
-COREGRAPHICS_EXPORT void CGContextSetLineWidth(CGContextRef context, float width);
+COREGRAPHICS_EXPORT void CGContextSetLineWidth(CGContextRef context, CGFloat width);
 COREGRAPHICS_EXPORT void CGContextSetLineCap(CGContextRef context, CGLineCap lineCap);
 COREGRAPHICS_EXPORT void CGContextSetLineJoin(CGContextRef context, CGLineJoin lineJoin);
-COREGRAPHICS_EXPORT void CGContextSetMiterLimit(CGContextRef context, float miterLimit);
-COREGRAPHICS_EXPORT void CGContextSetLineDash(CGContextRef context, float phase, const float *lengths, unsigned count);
+COREGRAPHICS_EXPORT void CGContextSetMiterLimit(CGContextRef context, CGFloat miterLimit);
+COREGRAPHICS_EXPORT void CGContextSetLineDash(CGContextRef context, CGFloat phase, const CGFloat *lengths, unsigned count);
 
 COREGRAPHICS_EXPORT void CGContextSetRenderingIntent(CGContextRef context, CGColorRenderingIntent renderingIntent);
 COREGRAPHICS_EXPORT void CGContextSetBlendMode(CGContextRef context, CGBlendMode blendMode);
 
-COREGRAPHICS_EXPORT void CGContextSetFlatness(CGContextRef context, float flatness);
+COREGRAPHICS_EXPORT void CGContextSetFlatness(CGContextRef context, CGFloat flatness);
 
 COREGRAPHICS_EXPORT void CGContextSetInterpolationQuality(CGContextRef context, CGInterpolationQuality quality);
 
-COREGRAPHICS_EXPORT void CGContextSetShadowWithColor(CGContextRef context, CGSize offset, float blur, CGColorRef color);
-COREGRAPHICS_EXPORT void CGContextSetShadow(CGContextRef context, CGSize offset, float blur);
+COREGRAPHICS_EXPORT void CGContextSetShadowWithColor(CGContextRef context, CGSize offset, CGFloat blur, CGColorRef color);
+COREGRAPHICS_EXPORT void CGContextSetShadow(CGContextRef context, CGSize offset, CGFloat blur);
 
 COREGRAPHICS_EXPORT void CGContextSetShouldAntialias(CGContextRef context, bool yesOrNo);
 
@@ -213,7 +213,7 @@ COREGRAPHICS_EXPORT void CGContextSetShouldAntialias(CGContextRef context, bool 
 COREGRAPHICS_EXPORT void CGContextStrokeLineSegments(CGContextRef context, const CGPoint *points, unsigned count);
 
 COREGRAPHICS_EXPORT void CGContextStrokeRect(CGContextRef context, CGRect rect);
-COREGRAPHICS_EXPORT void CGContextStrokeRectWithWidth(CGContextRef context, CGRect rect, float width);
+COREGRAPHICS_EXPORT void CGContextStrokeRectWithWidth(CGContextRef context, CGRect rect, CGFloat width);
 COREGRAPHICS_EXPORT void CGContextStrokeEllipseInRect(CGContextRef context, CGRect rect);
 
 COREGRAPHICS_EXPORT void CGContextFillRect(CGContextRef context, CGRect rect);
@@ -228,11 +228,11 @@ COREGRAPHICS_EXPORT void CGContextEOFillPath(CGContextRef context);
 COREGRAPHICS_EXPORT void CGContextClearRect(CGContextRef context, CGRect rect);
 
 COREGRAPHICS_EXPORT void CGContextShowGlyphs(CGContextRef context, const CGGlyph *glyphs, unsigned count);
-COREGRAPHICS_EXPORT void CGContextShowGlyphsAtPoint(CGContextRef context, float x, float y, const CGGlyph *glyphs, unsigned count);
+COREGRAPHICS_EXPORT void CGContextShowGlyphsAtPoint(CGContextRef context, CGFloat x, CGFloat y, const CGGlyph *glyphs, unsigned count);
 COREGRAPHICS_EXPORT void CGContextShowGlyphsWithAdvances(CGContextRef context, const CGGlyph *glyphs, const CGSize *advances, unsigned count);
 
 COREGRAPHICS_EXPORT void CGContextShowText(CGContextRef context, const char *text, unsigned count);
-COREGRAPHICS_EXPORT void CGContextShowTextAtPoint(CGContextRef context, float x, float y, const char *text, unsigned count);
+COREGRAPHICS_EXPORT void CGContextShowTextAtPoint(CGContextRef context, CGFloat x, CGFloat y, const char *text, unsigned count);
 
 COREGRAPHICS_EXPORT void CGContextDrawShading(CGContextRef context, CGShadingRef shading);
 COREGRAPHICS_EXPORT void CGContextDrawImage(CGContextRef context, CGRect rect, CGImageRef image);
